Teach node children iterator about element parity
NOTE: This is a performance, not correctness patch.

See discussion in the PR:
https://github.com/CAD97/sorbus/pull/1#issuecomment-612549153

On my machine, the flat iteration took an approximately 9% regression in time,
and about an 8% regression in throughput. On the other hand, the visitor saw
a 33% reduction in time, or a 50% improvement in throughput!

Inicola's machine (linux) saw a consistently directed but better result, with
approx. a 3/3% regression on flat iteration and a 43/75% (!!!) improvement for
the visitor benchmark.

Because the visitor benchmark is the longer-running and saw a large benefit,
in comparision to the smaller penalty on the flat iteration, along with the
assumption that visit is closer to a realistic workload, I have decided to
include this patch. It also feels "morally correct" to teach the children
iterator about the parity of the child elements, rather than checking the
alignment every single time.

As a Windows user with little performance measurement and tuning experience,
and with not-so-useful results from actually attempting to measure performance,
(surprise: all of the time is spent in `Children::fold`), I (CAD97) cannot do
much more in tuning here. I welcome those with more experience micro-optimizing
to take a stab at tuning iteration here, though to be frank, I do not expect
iteration of green syntax tree nodes to be a significant factor in performance
of any real application; the domain logic will likely easily outweigh it. Then,
the importance is more about how well the optimizer optimizes through the loop
than the actual operations and speed of the loop itself.