Make sure node dedupe checks all children's id eq
Pointed out by @steffahn on urlo.

Iterator::zip truncates the length of the longer iterator
and just uses the length of the shorter iterator.
In order to prevent nodes being deduped with nodes
that have a strict prefix of the same id children,
we have to first check the number of children present
separately from the Iterator::zip pair checking.

This is an isidiously hidden bug, as it requires a (partial)
hash collision before even revealing itself, as the nodes
would have to collide in the hash table before even being
checked for equality in this manner.