Only recurse GC if its possible children are free
As pointed out by @Darksonn

This initially always recursed into children,
to deal with the situation where a child node
is present more than once as a child of a node.
This is not an issue, however, as that node will
just be freed normally later as it reaches the
top of the processing stack.